Added support for custom exception handlers and created a defaule cascade one to use the error handling code that cascade has.
