+++ http/src/main/scala/com/paypal/cascade/http/server/SprayConfiguration.scala
@@ -15,7 +15,7 @@
  */
 package com.paypal.cascade.http.server
 
-import spray.routing.{RejectionHandler, Route}
+import spray.routing.{ExceptionHandler, RejectionHandler, Route}
 
 /**
  * This class provides configuration information for a spray service
@@ -24,11 +24,12 @@ class SprayConfiguration(val serviceName: String,
                          val port: Int,
                          val backlog: Int,
                          val route: Route,
+                         val customExceptionHandler: Option[ExceptionHandler] = None,
                          val customRejectionHandler: Option[RejectionHandler] = None)
 
 object SprayConfiguration {
-  def apply(serviceName: String, port: Int, backlog: Int, customRejectionHandler: Option[RejectionHandler] = None)
+  def apply(serviceName: String, port: Int, backlog: Int, customExceptionHandler: Option[ExceptionHandler] = None, customRejectionHandler: Option[RejectionHandler] = None)
            (route: Route): SprayConfiguration = {
-    new SprayConfiguration(serviceName, port, backlog, route, customRejectionHandler)
+    new SprayConfiguration(serviceName, port, backlog, route, customExceptionHandler, customRejectionHandler)
   }
 }
